強化

+++ Assets/Scripts/Save.cs
@@ -8,9 +8,11 @@ public class Save : MonoBehaviour {
     private int levelhash;//ハッシュ
     private int hpdata;//HP
     private int hphash;//ハッシュ
-    private int levelshash = 11;//ハッシュキーは11
+    private int levelshash = 107;//ハッシュキーは107
     private int hpshash = 23;
+    private int hphashhash;
     private int levelhashplus = 101;//強化
+    private int hpshashhash = 89;//ハッシュのハッシュキー
     public Datamanagement datamanagement;
     public string named;
 	public string Load(int num)//num番目のセーブデータを読み込む
@@ -21,29 +23,37 @@ public class Save : MonoBehaviour {
             {
                 if (PlayerPrefs.HasKey("name" + num))
                 {
-                    leveldata = PlayerPrefs.GetInt("level" + num);
-                    hpdata = PlayerPrefs.GetInt("hp" + num);
-                    levelhash = PlayerPrefs.GetInt("levelh" + num);
-                    hphash = PlayerPrefs.GetInt("hph" + num);
-                    if (leveldata % levelshash != levelhash)
+                    if (PlayerPrefs.HasKey("hphh" + num))
                     {
-                        PlayerPrefs.DeleteAll();//もし改変されていたらすべてのデータを削除する
-                        return "save data had breaked!";
+                        leveldata = PlayerPrefs.GetInt("level" + num);
+                        hpdata = PlayerPrefs.GetInt("hp" + num);
+                        levelhash = PlayerPrefs.GetInt("levelh" + num);
+                        hphash = PlayerPrefs.GetInt("hph" + num);
+                        hphashhash = PlayerPrefs.GetInt("hphh" + num);
+                        if (leveldata % levelshash != levelhash)
+                        {
+                            PlayerPrefs.DeleteAll();//もし改変されていたらすべてのデータを削除する
+                            return "save data had breaked!";
+                        }
+                        if (hphash % hpshashhash != hphashhash)
+                        {
+                            PlayerPrefs.DeleteAll();
+                        }
+                        if (hpdata % hpshash != hphash)
+                        {
+                            PlayerPrefs.DeleteAll();
+                            return "save data had breaked!";
+                        }
+                        if (leveldata % levelhashplus != 0)
+                        {
+                            PlayerPrefs.DeleteAll();
+                            return "save data had breaked!";
+                        }
+                        leveldata = leveldata % levelhashplus;
+                        datamanagement.Saver(hpdata, leveldata);
+                        named = PlayerPrefs.GetString("name" + num);
+                        return named;
                     }
-                    if (hpdata % hpshash != hphash)
-                    {
-                        PlayerPrefs.DeleteAll();
-                        return "save data had breaked!";
-                    }
-                    if (leveldata % levelhashplus != 0)
-                    {
-                        PlayerPrefs.DeleteAll();
-                        return "save data had breaked!";
-                    }
-                    leveldata = leveldata % levelhashplus;
-                    datamanagement.Saver(hpdata, leveldata);
-                    named = PlayerPrefs.GetString("name" + num);
-                    return named;
                 }
             }
         }
@@ -69,6 +79,7 @@ public class Save : MonoBehaviour {
         PlayerPrefs.SetInt("hp" + num, (datamanagement.hitpoint)*levelhashplus);
         int hphaser = (datamanagement.hitpoint * levelhashplus) % hpshash;
         PlayerPrefs.SetInt("hph" + num, hphaser);
+        PlayerPrefs.SetInt("hphh" + num, hphaser % hpshashhash);
         PlayerPrefs.SetInt("level" + num, (datamanagement.level));
         PlayerPrefs.SetInt("levelh" + num, ((datamanagement.level) % levelshash));
     }
